If there were no inline share policies, we add no resources

+++ backend/dataall/core/environment/services/env_share_policy_service.py
@@ -179,8 +179,9 @@ class SharePolicyService:
         new_resources.extend(SharePolicyService.get_resources_from_existing_inline_policy(
             AWSAccountId, iam_role_name, IAM_ACCESS_POINT_ROLE_POLICY
         ))
-        policy_doc["Statement"][0]["Resource"] = new_resources
-        SharePolicyService.remove_empty_statement(policy_doc)
+        if len(new_resources)>0:
+            policy_doc["Statement"][0]["Resource"] = new_resources
+            SharePolicyService.remove_empty_statement(policy_doc)
